backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.13
pydantic[email]==1.10.13
python-multipart==0.0.9
celery==5.3.6
redis==5.0.1
pymysql==1.1.0
bcrypt==4.1.2

